Moved dependencies to separate module
